+++ src/components/SharedDashboard/Messages.jsx
@@ -5,212 +5,39 @@ import { useLanguage } from '../../context/LanguageContext';
 import { FaPaperPlane, FaSearch, FaEllipsisV, FaPhone, FaVideo, FaComments, FaMicrophone, FaMicrophoneSlash, FaPhoneSlash } from 'react-icons/fa';
 import { toast } from 'react-hot-toast';
 import profile from '../../assets/profile.jpeg';
-import AgoraRTC from 'agora-rtc-sdk-ng';
 import { useTheme } from '../../context/ThemeContext';
 
 // Ringtone audio URL
-const RINGTONE_URL = '/ringtone.mp3'; // Ensure this file is in your public folder
-
-// Agora audio call hook (in this file for simplicity)
-function useAgoraAudioCall() {
-  const [inCall, setInCall] = useState(false);
-  const [isMuted, setIsMuted] = useState(false);
-  const [callError, setCallError] = useState(null);
-  const [incomingCall, setIncomingCall] = useState(null);
-  const clientRef = useRef(null);
-  const localAudioTrackRef = useRef(null);
-
-  const startCall = async ({ channelName, uid }) => {
-    setCallError(null);
-    clientRef.current = AgoraRTC.createClient({ mode: 'rtc', codec: 'vp8' });
-    try {
-      // Get token from backend
-      const res = await fetch('/api/agora/token', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ channelName, uid }),
-      });
-      const data = await res.json();
-
-      if (!data.token || !data.appId) {
-        throw new Error(data.error || 'Failed to get Agora token from backend');
-      }
-
-      await clientRef.current.join(data.appId, channelName, data.token, uid);
-      localAudioTrackRef.current = await AgoraRTC.createMicrophoneAudioTrack();
-      await clientRef.current.publish([localAudioTrackRef.current]);
-
-      // Listen for user joining (incoming call)
-      clientRef.current.on('user-joined', (user) => {
-        setIncomingCall({
-          uid: user.uid,
-          channelName,
-        });
-      });
-
-      clientRef.current.on('user-published', async (user, mediaType) => {
-        await clientRef.current.subscribe(user, mediaType);
-        if (mediaType === 'audio') {
-          user.audioTrack.play();
-        }
-      });
-
-      setInCall(true);
-    } catch (err) {
-      console.error('Error starting call:', err);
-      setCallError(err.message || 'Failed to start call');
-      setInCall(false);
-    }
-  };
-
-  const acceptCall = async () => {
-    if (!incomingCall) return;
-    try {
-      const { channelName, uid } = incomingCall;
-      await startCall({ channelName, uid });
-      setIncomingCall(null);
-    } catch (err) {
-      console.error('Error accepting call:', err);
-      setCallError(err.message || 'Failed to accept call');
-    }
-  };
-
-  const rejectCall = async () => {
-    if (!incomingCall) return;
-    try {
-      if (clientRef.current) {
-        await clientRef.current.leave();
-      }
-      setIncomingCall(null);
-    } catch (err) {
-      console.error('Error rejecting call:', err);
-    }
-  };
-
-  const leaveCall = async () => {
-    try {
-      if (localAudioTrackRef.current) {
-        localAudioTrackRef.current.close();
-      }
-      if (clientRef.current) {
-        await clientRef.current.leave();
-      }
-    } catch {}
-    setInCall(false);
-    setIsMuted(false);
-    setIncomingCall(null);
-  };
-
-  const toggleMute = async () => {
-    if (localAudioTrackRef.current) {
-      if (isMuted) {
-        await localAudioTrackRef.current.setEnabled(true);
-        setIsMuted(false);
-      } else {
-        await localAudioTrackRef.current.setEnabled(false);
-        setIsMuted(true);
-      }
-    }
-  };
-
-  return { inCall, startCall, leaveCall, isMuted, toggleMute, callError, incomingCall, acceptCall, rejectCall };
-}
+const RINGTONE_URL = '/ringtone.mp3';
 
 // Firestore Call State Management
 const CALLS_COLLECTION = 'calls';
 
-function useCallSignaling({ currentUser, otherUser }) {
-  const [callState, setCallState] = useState(null); // { id, status, caller, callee, channelName, startedAt, ... }
-  const [callDocId, setCallDocId] = useState(null);
-
-  // Listen for incoming/outgoing call state
-  useEffect(() => {
-    if (!currentUser) return;
-    const q = query(
-      collection(db, CALLS_COLLECTION),
-      where('participants', 'array-contains', currentUser.uid),
-      where('status', 'in', ['calling', 'ringing', 'active'])
-    );
-    const unsubscribe = onSnapshot(q, (snapshot) => {
-      if (!snapshot.empty) {
-        const callDoc = snapshot.docs[0];
-        setCallState({ id: callDoc.id, ...callDoc.data() });
-        setCallDocId(callDoc.id);
-      } else {
-        setCallState(null);
-        setCallDocId(null);
-      }
-    });
-    return () => unsubscribe();
-  }, [currentUser]);
-
-  // Initiate a call (caller)
-  const initiateCall = async () => {
-    if (!currentUser || !otherUser) return;
-    const channelName = [currentUser.uid, otherUser.id].sort().join('_');
-    const callDoc = {
-      participants: [currentUser.uid, otherUser.id],
-      caller: {
-        uid: currentUser.uid,
-        username: currentUser.displayName || currentUser.email,
-        avatarUrl: currentUser.photoURL || '',
-      },
-      callee: {
-        uid: otherUser.id,
-        username: otherUser.username,
-        avatarUrl: otherUser.avatarUrl || '',
-      },
-      channelName,
-      status: 'calling', // 'calling', 'ringing', 'active', 'ended', 'rejected', 'missed'
-      startedAt: serverTimestamp(),
-      acceptedAt: null,
-      endedAt: null,
-    };
-    const docRef = await addDoc(collection(db, CALLS_COLLECTION), callDoc);
-    setCallDocId(docRef.id);
-    setCallState({ id: docRef.id, ...callDoc });
-    return docRef.id;
+// Function to initiate a call (caller)
+const initiateCall = async (currentUser, otherUser) => {
+  if (!currentUser || !otherUser) return;
+  const channelName = [currentUser.uid, otherUser.id].sort().join('_');
+  const callDoc = {
+    participants: [currentUser.uid, otherUser.id],
+    caller: {
+      uid: currentUser.uid,
+      username: currentUser.displayName || currentUser.email,
+      avatarUrl: currentUser.photoURL || '',
+    },
+    callee: {
+      uid: otherUser.id,
+      username: otherUser.username,
+      avatarUrl: otherUser.avatarUrl || '',
+    },
+    channelName,
+    status: 'calling', // 'calling', 'ringing', 'active', 'ended', 'rejected', 'missed'
+    startedAt: serverTimestamp(),
+    acceptedAt: null,
+    endedAt: null,
   };
-
-  // Accept call (callee)
-  const acceptCall = async () => {
-    if (!callDocId) return;
-    await updateDoc(doc(db, CALLS_COLLECTION, callDocId), {
-      status: 'active',
-      acceptedAt: serverTimestamp(),
-    });
-  };
-
-  // Reject call (callee)
-  const rejectCall = async () => {
-    if (!callDocId) return;
-    await updateDoc(doc(db, CALLS_COLLECTION, callDocId), {
-      status: 'rejected',
-      endedAt: serverTimestamp(),
-    });
-  };
-
-  // End call (either side)
-  const endCall = async () => {
-    if (!callDocId) return;
-    await updateDoc(doc(db, CALLS_COLLECTION, callDocId), {
-      status: 'ended',
-      endedAt: serverTimestamp(),
-    });
-  };
-
-  // Cleanup call doc if ended
-  useEffect(() => {
-    if (callState && ['ended', 'rejected', 'missed'].includes(callState.status) && callDocId) {
-      // Optionally delete the call doc after a delay
-      setTimeout(() => {
-        deleteDoc(doc(db, CALLS_COLLECTION, callDocId));
-      }, 10000);
-    }
-  }, [callState, callDocId]);
-
-  return { callState, initiateCall, acceptCall, rejectCall, endCall };
-}
+  const docRef = await addDoc(collection(db, CALLS_COLLECTION), callDoc);
+  return docRef.id;
+};
 
 const Messages = () => {
   const [conversations, setConversations] = useState([]);
@@ -226,11 +53,8 @@ const Messages = () => {
   const messagesEndRef = useRef(null);
   const { t } = useLanguage();
   const { theme } = useTheme();
-  const { inCall, startCall, leaveCall, isMuted, toggleMute, callError, incomingCall, acceptCall, rejectCall } = useAgoraAudioCall();
-  const [callModalOpen, setCallModalOpen] = useState(false);
   const [loadingMessages, setLoadingMessages] = useState(false);
   const [typing, setTyping] = useState(false);
-  const ringtoneRef = useRef(null);
 
   // Fetch friend requests
   useEffect(() => {
@@ -482,132 +306,15 @@ const Messages = () => {
   // Find the other user in the selected chat
   const otherUser = selectedChat && users.find(u => u.id === selectedChat.participants.find(p => p !== auth.currentUser?.uid));
 
-  // --- Call Signaling Integration ---
-  const currentUser = auth.currentUser;
-  const {
-    callState,
-    initiateCall,
-    acceptCall: acceptSignalingCall,
-    rejectCall: rejectSignalingCall,
-    endCall: endSignalingCall
-  } = useCallSignaling({ currentUser, otherUser });
-
-  // --- Agora Audio Call Integration ---
-  const {
-    inCall: agoraInCall,
-    startCall: agoraStartCall,
-    leaveCall: agoraLeaveCall,
-    isMuted: agoraIsMuted,
-    toggleMute: agoraToggleMute,
-    callError: agoraCallError
-  } = useAgoraAudioCall();
-
-  // Speaker toggle state
-  const [isSpeaker, setIsSpeaker] = useState(false);
-  // Call duration state
-  const [callDuration, setCallDuration] = useState(0);
-  const callTimerRef = useRef(null);
-
-  // Start/stop call timer
-  useEffect(() => {
-    if (callState && callState.status === 'active') {
-      setCallDuration(0);
-      callTimerRef.current = setInterval(() => {
-        setCallDuration((d) => d + 1);
-      }, 1000);
-    } else {
-      clearInterval(callTimerRef.current);
-      setCallDuration(0);
+  // Handle initiating a call
+  const handleInitiateCall = async () => {
+    if (!otherUser) return;
+    try {
+      await initiateCall(auth.currentUser, otherUser);
+    } catch (error) {
+      console.error('Error initiating call:', error);
+      toast.error('Failed to initiate call');
     }
-    return () => clearInterval(callTimerRef.current);
-  }, [callState && callState.status === 'active']);
-
-  // Join/leave Agora channel based on call state
-  useEffect(() => {
-    if (callState && callState.status === 'active') {
-      // Only join if not already in call
-      if (!agoraInCall) {
-        agoraStartCall({ channelName: callState.channelName, uid: Number(currentUser.uid) });
-      }
-    } else if (agoraInCall && (!callState || ['ended', 'rejected', 'missed'].includes(callState.status))) {
-      agoraLeaveCall();
-    }
-    // eslint-disable-next-line
-  }, [callState && callState.status, agoraInCall]);
-
-  // Speaker toggle (if supported by Agora/browser)
-  const handleToggleSpeaker = () => {
-    // This is a placeholder: Agora Web SDK does not support speaker/earpiece switching on desktop browsers.
-    setIsSpeaker((s) => !s);
-    // Optionally, show a toast or info if not supported
-    toast('Speaker toggle is not supported on this device/browser.');
-  };
-
-  // Format call duration
-  const formatDuration = (seconds) => {
-    const m = Math.floor(seconds / 60).toString().padStart(2, '0');
-    const s = (seconds % 60).toString().padStart(2, '0');
-    return `${m}:${s}`;
-  };
-
-  // --- WhatsApp-like Call UI Components ---
-  const CallScreen = () => {
-    if (!callState) return null;
-    const isCaller = callState.caller.uid === currentUser.uid;
-    const peer = isCaller ? callState.callee : callState.caller;
-    // Outgoing call (caller)
-    if (callState.status === 'calling' && isCaller) {
-      return (
-        <div className="fixed inset-0 z-50 flex flex-col items-center justify-center bg-black bg-opacity-80">
-          <img src={peer.avatarUrl || profile} alt={peer.username} className="w-32 h-32 rounded-full mb-6 object-cover border-4 border-orange-500" />
-          <div className="text-2xl font-bold mb-2 text-white">{peer.username}</div>
-          <div className="text-lg text-gray-300 mb-8">Calling...</div>
-          <div className="flex gap-8">
-            <button onClick={endSignalingCall} className="w-16 h-16 rounded-full bg-red-600 flex items-center justify-center text-white text-3xl shadow-lg">
-              <FaPhoneSlash />
-            </button>
-          </div>
-        </div>
-      );
-    }
-    // Incoming call (callee)
-    if (callState.status === 'calling' && !isCaller) {
-      return (
-        <div className="fixed inset-0 z-50 flex flex-col items-center justify-center bg-black bg-opacity-80">
-          <img src={peer.avatarUrl || profile} alt={peer.username} className="w-32 h-32 rounded-full mb-6 object-cover border-4 border-orange-500" />
-          <div className="text-2xl font-bold mb-2 text-white">{peer.username}</div>
-          <div className="text-lg text-gray-300 mb-8">is calling you...</div>
-          <div className="flex gap-8">
-            <button onClick={acceptSignalingCall} className="w-16 h-16 rounded-full bg-green-500 flex items-center justify-center text-white text-3xl shadow-lg">
-              <FaPhone />
-            </button>
-            <button onClick={rejectSignalingCall} className="w-16 h-16 rounded-full bg-red-600 flex items-center justify-center text-white text-3xl shadow-lg">
-              <FaPhoneSlash />
-            </button>
-          </div>
-        </div>
-      );
-    }
-    // Active call (both sides)
-    if (callState.status === 'active') {
-      return (
-        <div className="fixed inset-0 z-50 flex flex-col items-center justify-center bg-black bg-opacity-80">
-          <img src={peer.avatarUrl || profile} alt={peer.username} className="w-32 h-32 rounded-full mb-6 object-cover border-4 border-orange-500" />
-          <div className="text-2xl font-bold mb-2 text-white">{peer.username}</div>
-          <div className="text-lg text-gray-300 mb-2">{formatDuration(callDuration)}</div>
-          <div className="flex gap-8 mt-6">
-            <button onClick={agoraToggleMute} className={`w-16 h-16 rounded-full flex items-center justify-center text-3xl shadow-lg ${agoraIsMuted ? 'bg-gray-400 text-white' : 'bg-orange-500 text-white'}`}>{agoraIsMuted ? <FaMicrophoneSlash /> : <FaMicrophone />}</button>
-            <button onClick={handleToggleSpeaker} className={`w-16 h-16 rounded-full flex items-center justify-center text-3xl shadow-lg ${isSpeaker ? 'bg-orange-400 text-white' : 'bg-gray-300 text-gray-700'}`}>🔊</button>
-            <button onClick={endSignalingCall} className="w-16 h-16 rounded-full bg-red-600 flex items-center justify-center text-white text-3xl shadow-lg">
-              <FaPhoneSlash />
-            </button>
-          </div>
-          {agoraCallError && <div className="text-red-400 mt-4">{agoraCallError}</div>}
-        </div>
-      );
-    }
-    // Ended/rejected/missed: no overlay
-    return null;
   };
 
   // --- Friend Request Modal ---
@@ -658,28 +365,6 @@ const Messages = () => {
       </div>
     </div>
   );
-
-  // Play/stop ringtone for incoming calls
-  useEffect(() => {
-    if (!ringtoneRef.current) {
-      ringtoneRef.current = new Audio(RINGTONE_URL);
-      ringtoneRef.current.loop = true;
-    }
-    // Play ringtone only if incoming call, status is 'calling', and not in an active call
-    const isIncoming = callState && callState.status === 'calling' && callState.callee && callState.callee.uid === currentUser?.uid && !agoraInCall;
-    if (isIncoming) {
-      ringtoneRef.current.play().catch(() => {});
-    } else {
-      ringtoneRef.current.pause();
-      ringtoneRef.current.currentTime = 0;
-    }
-    return () => {
-      if (ringtoneRef.current) {
-        ringtoneRef.current.pause();
-        ringtoneRef.current.currentTime = 0;
-      }
-    };
-  }, [callState?.status, callState?.callee?.uid, currentUser?.uid, agoraInCall]);
 
   // --- Friend Requests List ---
   const FriendRequestsList = () => (
@@ -732,8 +417,6 @@ const Messages = () => {
 
   return (
     <div className={`flex h-[calc(100vh-200px)] rounded-lg shadow-lg overflow-hidden ${theme === 'dark' ? 'bg-gray-900 text-gray-100' : 'bg-white'}`}>
-      {/* WhatsApp-like Call Overlay */}
-      {callState && ['calling', 'active'].includes(callState.status) && <CallScreen />}
       {/* Friend Requests Badge */}
       {friendRequests.length > 0 && (
         <div className="absolute top-2 right-2 bg-red-500 text-white rounded-full w-6 h-6 flex items-center justify-center text-sm font-medium shadow-lg animate-bounce">
@@ -852,7 +535,7 @@ const Messages = () => {
                 </div>
               </div>
               <div className="flex items-center space-x-4">
-                <button className={`p-2 transition-colors ${theme === 'dark' ? 'text-gray-400 hover:text-[#FFD966]' : 'text-gray-600 hover:text-orange-500'}`} onClick={initiateCall} disabled={inCall} title="Start Audio Call">
+                <button className={`p-2 transition-colors ${theme === 'dark' ? 'text-gray-400 hover:text-[#FFD966]' : 'text-gray-600 hover:text-orange-500'}`} onClick={handleInitiateCall} title="Start Audio Call">
                   <FaPhone />
                 </button>
                 <button className={`p-2 transition-colors ${theme === 'dark' ? 'text-gray-400 hover:text-[#FFD966]' : 'text-gray-600 hover:text-orange-500'}`}>
@@ -1039,4 +722,4 @@ const Messages = () => {
   );
 };
 
-export default Messages; 
+export default Messages;
